PyTorch Update CNN
I got rid of a lot of the over-fitting code as I thought it might be wrong to heavily modify the Training a Classifier script on PyTorch's website when we call this a re annotated version. I left in the text on over-fitting though. Fixed a couple of grammatical errors too, lol.
ex. 
a colon must be preceded by a complete sentence.
Anyways, this one was fun and I should be done with the next one tomorrow.
Email soon!
- Sebastian